backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["PomPerWoW <chatthanes22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "^3.0.3"
flask-smorest = "^0.44.0"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.1"
pymysql = "^1.1.1"
cryptography = "^43.0.1"
marshmallow = "^3.22.0"
flask-cors = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
